+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Reload report after threshold modifications
 - Highlight selected threshold row in sidebar
 ### Changed
-- Combine column, sorting and other dimension options in a single dialog
+- Consolidate column, sort, Top N and time aggregation settings in a grid layout
 
 ## 5.6.1 - 2025-06-09
 ### Fixed